Some Movement and Poofs and Tiles - Oh my!
Added 5 seperate scripts, only 4 of these are functional. The
PassiveMover script is currently completely inoperable. However the
MovementScript, PoofScript, and TileScript interact as they should.
Renamed the tiles, and in essence reworked the coordinate system of the
game. If you want the poofs to appear to be moving along the lines of
the tiles, you can edit the center position of the poofs from the sprite
definitions in the editor.