Can now "create" an account by posting to /create
Sending a post request to the /create path will create a row in
the SQL database. The post body must have a 'name', 'username,
and 'email' parameter in it.

Also commented out importing user_building in most cases as it was
causing errors. Fixed other minor errors with SQLalchemy syntax