Logging in now checks a users password
When you log in via the server It checks the password to see if it
is the correct one.

I had to update the User module to add the password column.

I also updated the tables so that username is unique to reflect
that that is what we are using to log in right now.