Logging in now returns user information
Logging in will now return the user's information via JSON.

There is a bug that if you try and create a duplicate user and
then login to the server an exception will be thrown. It seems
the session will need to be rolled back before it can be used
again.