Logging in now returns a users building data
When you log in the server should return the user info as well as
info about the users buildings. The JSON contains two arrays, one
for the users resource buildings and one for the users decorative
buildings. There is still some information missing from the JSON
That I need to get from the building info table.

Miscellaneous things:
* Moved queries back out.
* Edited the tables so that they is more consistent