Saving now saves buildings
When saving you need to add a resource_buildings array and a
decorative_buildings array. These will either save new buildings
or update old ones depending on if the "new" field is "true" or
not.

Saving POST request format:

{
	"user_id" : "1",
	"username" : "ted1",
	"name" : "ted",
	"email" : "poop@poop.com",
	"level" : "2",
	"password" : "password",
	"experience" : "25",
	"hq_level" : "14",
	"resource_buildings" : [
	{
		"new" : "false",
		"id" : "1",
		"position_x" : "1",
		"position_y" : "1",
	},
	{
		"new" : "true",
		"position_x" : "13",
		"position_y" : "13",
		"building_info_id" : "1"
	}

	],
	"decorative_buildings" : [
	{
		"new" : "false",
		"id" : "1",
		"position_x" : "2",
		"position_y" : "2",
		"level" : "4"
	},

		"new" : "true",
		"position_x" : "5",
		"position_y" : "5",
		"level" : "1",
		"building_info_id" : "1"
	}
	]
}

When you are updating a building you need to include the "id"
field to identify which building to update. When creating a
building you will need to include the "building_info_id" field to
link a building to its type.