Randomly assigns a poof an "onTile"
Poofs that are currently in the scene will eventually need to be
spawned/generated like the game grid. If poofs already exist in the
scene, we cannot give them a start tile which forces us to calculate a
random spot for them/calculate where they are and assign the nearest
tile. Regardless, poofs won't start on an edge on game startup which
will force them to walk through a grid square to get to their "onTile."