Added comments to describe future goals for tile script
In the future, we are going to need the entire scene to be generated
from tilescript to avoid any race conditions. In otherwords, everything
in our scene hinges on the existence of a grid. We can generate the
grid, then load our save data, then populate the grid.

I also added a small note for save data. If we can't put together the
front and back ends before the demo, we can load a predefined save state
from a JSON file.