Resources are now incremented every second
The Resource building class will work for all resource buildings.
It has a timer that will tick once a second and then you can call
the ResourceGain method on the resource incrementer. The
incrementer is now a singleton like the SaveState.

Future:
    We need a field on the resource building that is its resources
generated per second. We will need to load this from the server.
We will also need a max resources that the player can hold at a
time and set the slider value to current / max * 100.  The
resource building class needs to be more fleshed out right now
it only produces gold. This also applies to the Incrementer.

issues:
* The slider doesn't update when its value updates...
* The timer goes after play has stopped.