Data pulled from server is now stored in SaveState
all of the fields in SaveState are now updated with the info from
the server. We still need to populate the game grid with these
values. Also the server needs to be updated to reflect the user
info we are now storing.

Miscellaneous:
* The BuildingManager is now instantiated at game start.
* The Make file now builds AND populates the database one command.

Notes:
The way we create buildings doesn't really read from the server
that well. We may also want to move the pullFromServer and
pushToServer methods into a general server like class?