Scene now loads from JSON
Objects spawn but we have a couple bugs:
1. Occasionally starting the game causes it to freeze.
2. Objects spawn "under the grid" - I'm not exactly sure what is
happening here because trees will get uncovered if you hover over them.
3. Tiles don't receive any data about objects being on top of them. So
while I can't place a tree on a tree, the tiles do not turn red for
loaded trees.
4. Objects spawn on their exactly on top of the tile by which they are
referenced. It seems we want 4x4 objects to sit centered among those
tiles.

I expanded the grid to be 15x15 because the populate script generates
buildings that are outside the range of our grid.