Further broke the building menu
Tried to move away from an index based building function. Implemented a
string based function, and it failed. This is because buttons and
buildings aren't related in name (also strings are just asking for
trouble). I

I think a good way to resolve this would be to remove the distinction
between buttons and buildings. Instead, buildings should be buttons.
Buildings could benefit from the drag functionality anyhow. However, the
issue would be dealing with the building in the menu thinking it was an
entity on the grid. We could have some strange behavior.