Issue with moving building
Buildings seem to be moved with ease. However, we're running into an
issue with the dictionary where buildings aren't being re-added
correctly to the dictionary. If you move a building away then try to
move it again, you'll receive an error about a missing tuple in the
dictionary. Try moving it back to the second position and it will get
stuck. This obviously effects delete.